+++ pygplib/grst.py
@@ -1,6 +1,7 @@
 """Graph relational structure class"""
 
 import functools
+from typing import overload
 
 from .name     import NameMgr
 from .symrelst import SymRelSt
@@ -33,13 +34,12 @@ class GrSt(SymRelSt):
         """Initialize a graph structure.
 
         Args:
-            vertex_list: list of vertices
-            edge_list: list of edges (pairs of vertices)
+            vertex_list: list of distinct vertices in an arbitrary order
+            edge_list: list of distinct edges (pairs of vertices) in an
+            arbitrary order
             encoding: encoding type ("edge", "clique", "direct", "log")
             prefix: prefix of vertex name, which must be uppercase.
         """
-        if len(set(vertex_list)) != len(vertex_list):
-            raise Exception(f"duplicate vertex found: {vertex_list}")
         for edge in edge_list:
             if len(edge) != 2:
                 raise Exception(f"edge must be of size 2: {edge}")
@@ -48,94 +48,183 @@ class GrSt(SymRelSt):
             if edge[0] not in vertex_list\
                 or edge[1] not in vertex_list:
                 raise Exception(f"invalid vertex found: {edge}")
+        if encoding not in {"edge", "clique", "direct"}:
+            raise Exception(f"unsupported encoding type: {encoding}")
         if not (prefix.isalpha() and prefix.isupper()):
-            raise ValueError(\
+            raise Exception(\
                 "All characters must be alphabetic and uppercase letters")
 
         self._encoding = encoding
-        self._prefix = prefix
+        """encoding type"""
+        self._prefix   = prefix
+        """prefix of vertex name"""
+        for vertex in vertex_list:
+            NameMgr.lookup_index(self._prefix + f"{vertex}")
+        self._verts = tuple(vertex_list)
+        """vertices"""
+        self._edges = tuple([tuple(sorted(edge)) for edge in edge_list])
+        """edges"""
+        if len(set(self._verts)) != len(vertex_list):
+            raise Exception(f"duplicate vertex found: {vertex_list}")
+        if len(set(self._edges)) != len(edge_list):
+            raise Exception(f"duplicate edge found: {edge_list}")
+        # vertex_to_object() and object_to_vertex() are now available!
 
-        self.verts = \
-            tuple([NameMgr.lookup_index(self._prefix + f"{vertex}") \
-                                        for vertex in set(vertex_list)])
-        self.edges = tuple(self._normalize_edge_list(edge_list))
-
+        objects = self.vertex_to_object(self._verts)
         if self._encoding == "edge":
-            super().__init__(self.verts, self.edges)
+            relation = self.vertex_to_object(self._edges)
         elif self._encoding == "clique":
-            super().__init__(self.verts, \
-                Ecc(self.verts, self.edges).compute_separating_ecc())
+            ecc = Ecc(self._verts, self._edges).compute_separating_ecc()
+            relation = self.vertex_to_object(ecc)
         elif self._encoding == "direct":
-            super().__init__(self.verts, \
-                tuple([(i+1,) for i in range(len(self.verts))]) \
-                )
+            relation = self.vertex_to_object(tuple([(v,) for v in self._verts]))
         elif self._encoding == "log":
             raise Exception("not-yet-implemented")
         else:
             raise Exception(f"invalid encoding type: {self._encoding}")
+        super().__init__(objects, relation)
 
-    def _normalize_edge_list(self, edge_list: list) -> list:
-        res = []
-        for edge in edge_list:
-            i = self.vertex_to_object(edge[0])
-            j = self.vertex_to_object(edge[1])
-            normalized = tuple(sorted([i,j]))
-            if normalized not in res:
-                res.append(normalized)
-        return res
+    def vertex_to_object_int(self, vertex: int) -> int:
+        """Converts vertex to object (constant symbol index).
 
-    def vertex_to_object(self, vertex: int) -> int:
+        Args:
+            tup: tuple of vertices
+        Returns:
+            tuple of domain objects (constant symbol indices)
+        """
+        if vertex not in self._verts:
+            raise Exception(f"invalid vertex: {vertex}")
         name = self._prefix + f"{vertex}" 
         if not NameMgr.has_index(name):
-            raise Exception(f"invalid vertex: {vertex}")
-        return NameMgr.lookup_index(name)
+            raise Exception(f"vertex: {vertex} not yet registered to NameMgr.")
+        obj = NameMgr.lookup_index(name)
+        return obj
 
-    def object_to_vertex(self, i: int) -> int:
-        if not NameMgr.has_name(i):
-            raise Exception(f"invalid symbol: {i}")
-        if not NameMgr.is_constant(i):
-            raise Exception(f"invalid symbol: {i}")
+    def vertex_to_object_tuple(self, tup: tuple[int]) -> tuple[int]:
+        """Converts vertex to object (constant symbol index).
 
-        name = NameMgr.lookup_name(i)
+        Args:
+            tup: tuple of (tuples of) vertices
+        Returns:
+            tuple of (tuples of) domain objects (constant symbol indices)
+        """
+        if not isinstance(tup, tuple):
+            raise TypeError
+        return tuple([self.vertex_to_object_int(x) if isinstance(x, int)\
+                        else self.vertex_to_object_tuple(x)\
+                        for x in tup])
+
+    @overload
+    def vertex_to_object(self, x: int) -> int:
+        pass
+
+    @overload
+    def vertex_to_object(self, x: tuple) -> tuple:
+        pass
+
+    def vertex_to_object(self, x):
+        """Converts vertex to object (constant symbol index).
+        
+        Alias of vertex_to_object_int() and vertex_to_object_tuple()
+        """
+        res = None
+        if isinstance(x, int):
+            res = self.vertex_to_object_int(x)
+        else:
+            res = self.vertex_to_object_tuple(x)
+        return res
+
+    def object_to_vertex(self, obj: int) -> int:
+        """Converts object (constant symbol index) to vertex.
+
+        Args:
+            obj: domain object (constant symbol index)
+        Returns:
+            index of vertex.
+        """
+        name = NameMgr.lookup_name(obj)
         if not (name.find(self._prefix) == 0 \
             and name[len(self._prefix) :].isdigit()):
-            raise ValueError(f"{NameMgr.lookup_name(i)} is not vertex name")
+            raise Exception(f"{name} is not vertex name")
 
         return int(name[len(self._prefix) :])
 
     def adjacent(self, i: int, j: int) -> bool:
-        if self._encoding in {"edge", "clique"}:
-            return True in [k in self.get_code(j) for k in self.get_code(i)]\
-                    and not self.equal(i,j)
-        if self._encoding == "direct":
-            u = self.object_to_vertex(i)
-            v = self.object_to_vertex(j)
-            return tuple(sorted([u,v])) in self.edges
-        else:
-            raise Exception(\
-                f"encoding type {self._encoding} does not match with this method")
+        """Determines if constants (meaning vertices) are adjacent.
+
+        Args:
+            i: constant symbol index (meaning vertex)
+            j: constant symbol index (meaning vertex)
+        Returns:
+            True if adjacent, and False otherwise.
+        """
+        if i not in self.domain:
+            raise Exception(f"{i} is not a domain object.")
+        if j not in self.domain:
+            raise Exception(f"{j} is not a domain object.")
+        u = self.object_to_vertex(i)
+        v = self.object_to_vertex(j)
+        return tuple(sorted([u,v])) in self._edges
 
     def equal(self, i: int, j: int) -> bool:
-        return self.get_code(i) == self.get_code(j)
+        """Determines if constants (meaning vertices) are equal with each other.
 
-    def _get_lit_list(self, i: int) -> list[Prop]:
-        if NameMgr.is_constant(i):
-            return [Prop.true_const() if pos+1 in self.get_code(i) \
+        Args:
+            i: constant symbol index (meaning vertex)
+            j: constant symbol index (meaning vertex)
+        Returns:
+            True if equal, and False otherwise.
+        """
+        if i not in self.domain:
+            raise Exception(f"{i} is not a domain object.")
+        if j not in self.domain:
+            raise Exception(f"{j} is not a domain object.")
+        return i == j
+
+    def _get_lit_list(self, index: int) -> list[Prop]:
+        """Gets list of literals, given a symbol index.
+
+        Args:
+            index: symbol index (constant symbol or variable symbol)
+        Returns:
+            list of literals (Boolean variables' objects of Prop class)
+        """
+        if index in self.domain:
+            return [Prop.true_const() if pos+1 in self.get_code(index) \
                         else Prop.false_const() \
                             for pos in range(self.code_length)]
         else:
-            return [Prop.var(j) for j in self.get_boolean_var_list(i)]
+            return [Prop.var(j) for j in self.get_boolean_var_list(index)]
 
     def encode_eq(self, i: int, j: int) -> Prop:
+        """Encodes predicate of equality relation, given two symbols.
+
+        Args:
+            i: symbol index (constant symbol or variable symbol)
+            j: symbol index (constant symbol or variable symbol)
+        Returns:
+            formula object of Prop class
+        """
+        res = None
         li = Prop.bitwise_binop(Prop.get_iff_tag(), \
                                 self._get_lit_list(i), \
                                 self._get_lit_list(j))
         if Prop.bipartite_order:
-            return Prop.binop_batch(Prop.get_land_tag(), li)
+            res = Prop.binop_batch(Prop.get_land_tag(), li)
         else:
-            return functools.reduce(lambda x,y: Prop.land(x,y), li)
+            res = functools.reduce(lambda x,y: Prop.land(x,y), li)
+        return res
 
     def encode_edg(self, i: int, j: int) -> Prop:
+        """Encodes predicate of adjacency relation, given two symbols.
+
+        Args:
+            i: symbol index (constant symbol or variable symbol)
+            j: symbol index (constant symbol or variable symbol)
+        Returns:
+            formula object of Prop class
+        """
+        res = None
         if self._encoding in {"edge", "clique"}:
             li = Prop.bitwise_binop(Prop.get_land_tag(), \
                                     self._get_lit_list(i), \
@@ -144,49 +233,57 @@ class GrSt(SymRelSt):
                 res = Prop.binop_batch(Prop.get_lor_tag(), li)
             else:
                 res = functools.reduce(lambda x,y: Prop.lor(x,y), li)
-            return Prop.land(res, Prop.neg(self.encode_eq(i,j)))
+            res = Prop.land(res, Prop.neg(self.encode_eq(i,j)))
         elif self._encoding == "direct":
             lits = [self._get_lit_list(i),self._get_lit_list(j)]
             li = [Prop.lor(\
-                    Prop.land(lits[0][self._object_to_pos(e[0])], \
-                        lits[1][self._object_to_pos(e[1])]),\
-                    Prop.land(lits[0][self._object_to_pos(e[1])], \
-                        lits[1][self._object_to_pos(e[0])]),\
-                        ) for e in self.edges]
+                    Prop.land(\
+                        lits[0][self._object_to_pos(edge[0])],\
+                        lits[1][self._object_to_pos(edge[1])]),\
+                    Prop.land(\
+                        lits[0][self._object_to_pos(edge[1])],\
+                        lits[1][self._object_to_pos(edge[0])]),\
+                        )\
+                        for edge in self.vertex_to_object(self._edges)]
             if Prop.bipartite_order:
-                return Prop.binop_batch(Prop.get_lor_tag(), li)
+                res = Prop.binop_batch(Prop.get_lor_tag(), li)
             else:
-                return functools.reduce(lambda x,y: Prop.lor(x,y), li)
+                res = functools.reduce(lambda x,y: Prop.lor(x,y), li)
         else:
-            raise Exception(\
-                f"encoding type {self._encoding} does not match with this method")
+            raise Exception(f"not yet implemented")
+        return res
 
-    def encode_T(self) -> Prop:
+    def encode_T(self) -> Prop: 
+        """Encodes true constant object of Fog class to Prop object."""
         return Prop.true_const()
 
     def encode_F(self) -> Prop:
+        """Encodes false constant object of Fog class to Prop object."""
         return Prop.false_const()
 
-    def compute_domain_constraint(self, var: int):
+    def compute_domain_constraint(self, var: int) -> Prop:
+        """Computes domain constraint for a first-order variable.
+
+        Args:
+            var: index of a first-order variable symbol
+        Returns:
+            formula object of Prop class
+        """
+        if not NameMgr.is_variable(var):
+            raise Exception(f"symbol index {var} is not a variable symbol")
+        res = None
         if self._encoding in ["edge", "clique"]:
-            return self._compute_domain_constraint_DNF_clique_encoding(var)
+            res = self._compute_domain_constraint_DNF(var)
         elif self._encoding == "direct":
-            return self._compute_domain_constraint_direct_encoding(var)
-        elif self._encoding == "log":
-            raise Exception
+            res = self._compute_domain_constraint_direct_encoding(var)
         else:
-            raise Exception
-            
-    def _compute_domain_constraint_DNF_clique_encoding(self, index: int)\
+            raise Exception(f"NotImplementedError")
+        return res
+
+    def _compute_domain_constraint_DNF(self, index: int)\
         -> Prop:
         """Computes a constraint in DNF for a given first-order variable
         such that the variable runs over the domain of this structure.
-
-        Note:
-            Use this for edge-encoding or clique-encoding.
-
-        Raises:
-            Exception: if index is not a variable symbol.
 
         Args:
             index: a first-order variable index
@@ -194,11 +291,6 @@ class GrSt(SymRelSt):
         Returns:
             Prop: Prop class object of the computed DNF constraint.
         """
-        if self._encoding not in {"edge", "clique"}:
-            raise Exception(\
-            f"encoding type {self._encoding} does not match with this method")
-        if not NameMgr.is_variable(index):
-            raise Exception(f"symbol index {index} is not a variable symbol")
         dnf = []
         for obj in self.domain:
             lits = self._get_lit_list(index)
@@ -216,12 +308,18 @@ class GrSt(SymRelSt):
         else:
             return functools.reduce(lambda a, b: Prop.lor(a, b), dnf)
 
-    def _compute_domain_constraint_direct_encoding(self, index: int):
+    def _compute_domain_constraint_direct_encoding(self, index: int) -> Prop:
+        """Computes domain constraint for a first-order variable
+        (direct-encoding version).
+
+        Args:
+            var: index of a first-order variable symbol
+        Returns:
+            formula object of Prop class
+        """
         if self._encoding != "direct":
             raise Exception(\
             f"encoding type {self._encoding} does not match with this method")
-        if not NameMgr.is_variable(index):
-            raise Exception(f"symbol index {index} is not a variable symbol")
 
         lits = self._get_lit_list(index)
         # at least one constraint